Party Codes

Party Codes

party_code

Format

A data frame with two variables:

party_code

A 1-3 character alphabetic code assigned to a party by the Federal Election Commission. The party code for a specific party always remains the same.

party

Official name of party as recognized by the FEC.

For further details and source, see http://www.fec.gov/finance/disclosure/metadata/DataDictionaryPartyCodeDescriptions.shtml